move mapping funcs to mapping.rs
