+++ src/meta/src/stream/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod mapping;
 mod scale;
 mod scheduler;
 mod source_manager;
@@ -20,6 +21,7 @@ mod stream_manager;
 #[cfg(test)]
 mod test_fragmenter;
 
+pub use mapping::*;
 pub use scale::*;
 pub use scheduler::*;
 pub use source_manager::*;
