+++ src/meta/src/stream/scheduler.rs
@@ -19,8 +19,7 @@ use anyhow::{anyhow, Context};
 use itertools::Itertools;
 use rand::prelude::SliceRandom;
 use risingwave_common::bail;
-use risingwave_common::buffer::{Bitmap, BitmapBuilder};
-use risingwave_common::types::{ParallelUnitId, VnodeMapping, VIRTUAL_NODE_COUNT};
+use risingwave_common::types::VnodeMapping;
 use risingwave_common::util::compress::compress_data;
 use risingwave_pb::common::{ActorInfo, ParallelUnit, ParallelUnitMapping, WorkerNode};
 use risingwave_pb::meta::table_fragments::fragment::FragmentDistributionType;
@@ -28,6 +27,7 @@ use risingwave_pb::meta::table_fragments::Fragment;
 
 use crate::manager::{WorkerId, WorkerLocations};
 use crate::model::ActorId;
+use crate::stream::{build_vnode_mapping, vnode_mapping_to_bitmaps};
 use crate::MetaResult;
 
 /// [`Scheduler`] defines schedule logic for mv actors.
@@ -218,7 +218,6 @@ impl Scheduler {
 
             // Build vnode mapping according to the parallel units.
             let vnode_mapping = self.set_fragment_vnode_mapping(fragment, &parallel_units)?;
-
             let vnode_bitmaps = vnode_mapping_to_bitmaps(vnode_mapping);
 
             // Record actor locations and set vnodes into the actors.
@@ -260,50 +259,6 @@ impl Scheduler {
 
         Ok(vnode_mapping)
     }
-}
-
-pub(crate) fn vnode_mapping_to_bitmaps(
-    vnode_mapping: VnodeMapping,
-) -> HashMap<ParallelUnitId, Bitmap> {
-    let mut vnode_bitmaps = HashMap::new();
-    vnode_mapping
-        .iter()
-        .enumerate()
-        .for_each(|(vnode, parallel_unit)| {
-            vnode_bitmaps
-                .entry(*parallel_unit)
-                .or_insert_with(|| BitmapBuilder::zeroed(VIRTUAL_NODE_COUNT))
-                .set(vnode, true);
-        });
-    vnode_bitmaps
-        .into_iter()
-        .map(|(u, b)| (u, b.finish()))
-        .collect()
-}
-
-/// Build a vnode mapping according to parallel units where the fragment is scheduled.
-/// For example, if `parallel_units` is `[0, 1, 2]`, and the total vnode count is 10, we'll
-/// generate mapping like `[0, 0, 0, 0, 1, 1, 1, 2, 2, 2]`.
-pub(crate) fn build_vnode_mapping(parallel_units: &[ParallelUnit]) -> VnodeMapping {
-    let mut vnode_mapping = Vec::with_capacity(VIRTUAL_NODE_COUNT);
-
-    let hash_shard_size = VIRTUAL_NODE_COUNT / parallel_units.len();
-    let mut one_more_count = VIRTUAL_NODE_COUNT % parallel_units.len();
-    let mut init_bound = 0;
-
-    parallel_units.iter().for_each(|parallel_unit| {
-        let vnode_count = if one_more_count > 0 {
-            one_more_count -= 1;
-            hash_shard_size + 1
-        } else {
-            hash_shard_size
-        };
-        let parallel_unit_id = parallel_unit.id;
-        init_bound += vnode_count;
-        vnode_mapping.resize(init_bound, parallel_unit_id);
-    });
-
-    vnode_mapping
 }
 
 #[cfg(test)]
